Disable optimizations in the test suites
Introduction of the parameter validation callbacks implemented in terms
of setjmp/longjmp causes certain tests compilation to raise warning
about clobbering certain variables. This happens when the are not
allocated on the stack. Disabling the compiler optimizations prevents
the warnings from being raised. Note that the `-O0` option is present in
the Makefile, whereas it is not used when building with CMake.